ml_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch
torchvision
Pillow
numpy
pandas
scikit-learn

